+++ pkg/front_end/lib/src/source/source_compilation_unit.dart
@@ -5,6 +5,8 @@
 import 'package:_fe_analyzer_shared/src/parser/class_member_parser.dart'
     show ClassMemberParser;
 import 'package:_fe_analyzer_shared/src/scanner/scanner.dart' show Token;
+import 'package:_fe_analyzer_shared/src/util/libraries_specification.dart'
+    show Importability;
 import 'package:kernel/ast.dart' hide Combinator, MapLiteralEntry;
 import 'package:kernel/reference_from_index.dart' show IndexedLibrary;
 
@@ -167,7 +169,10 @@ class SourceCompilationUnitImpl implements SourceCompilationUnit {
   final bool isAugmenting;
 
   @override
-  final bool isUnsupported;
+  final bool conditionalImportSupported;
+
+  @override
+  final Importability importability;
 
   late final LookupScope _compilationUnitScope;
 
@@ -194,7 +199,8 @@ class SourceCompilationUnitImpl implements SourceCompilationUnit {
     required bool? referenceIsPartOwner,
     required bool forPatchLibrary,
     required bool isAugmenting,
-    required bool isUnsupported,
+    required bool conditionalImportSupported,
+    required Importability importability,
     required SourceLoader loader,
     required bool mayImplementRestrictedTypes,
   }) {
@@ -220,7 +226,8 @@ class SourceCompilationUnitImpl implements SourceCompilationUnit {
       referenceIsPartOwner: referenceIsPartOwner,
       forPatchLibrary: forPatchLibrary,
       isAugmenting: isAugmenting,
-      isUnsupported: isUnsupported,
+      conditionalImportSupported: conditionalImportSupported,
+      importability: importability,
       loader: loader,
       mayImplementRestrictedTypes: mayImplementRestrictedTypes,
     );
@@ -244,7 +251,8 @@ class SourceCompilationUnitImpl implements SourceCompilationUnit {
     required bool? referenceIsPartOwner,
     required this.forPatchLibrary,
     required this.isAugmenting,
-    required this.isUnsupported,
+    required this.conditionalImportSupported,
+    required this.importability,
     required this.loader,
     required this.mayImplementRestrictedTypes,
   }) : _languageVersion = packageLanguageVersion,
@@ -657,7 +665,7 @@ class SourceCompilationUnitImpl implements SourceCompilationUnit {
           target: library,
           indexedLibrary: indexedLibrary,
           referenceIsPartOwner: _referenceIsPartOwner,
-          isUnsupported: isUnsupported,
+          conditionalImportSupported: conditionalImportSupported,
           isAugmentation: forAugmentationLibrary,
           isPatch: forPatchLibrary,
           parentScope: _parentScope,