+++ runtime/vm/compiler/backend/il_arm64.cc
@@ -1921,8 +1921,8 @@ void LoadIndexedInstr::EmitNativeCode(FlowGraphCompiler* compiler) {
   auto const rep =
       RepresentationUtils::RepresentationOfArrayElement(class_id());
 
-  if (!compiler->is_optimizing() && FLAG_target_thread_sanitizer) {
-    EmitTsanCallUnopt(compiler, this, [&]() -> const RuntimeEntry& {
+  if (!compiler->is_optimizing() && sanitize()) {
+    EmitSanCallUnopt(compiler, this, [&]() -> const RuntimeEntry& {
       if (index.IsRegister()) {
         __ ComputeElementAddressForRegIndex(R0, IsUntagged(), class_id(),
                                             index_scale(), index_unboxed_,
@@ -1934,15 +1934,15 @@ void LoadIndexedInstr::EmitNativeCode(FlowGraphCompiler* compiler) {
       }
       switch (RepresentationUtils::ValueSize(rep)) {
         case 1:
-          return kTsanRead1RuntimeEntry;
+          return kSanRead1RuntimeEntry;
         case 2:
-          return kTsanRead2RuntimeEntry;
+          return kSanRead2RuntimeEntry;
         case 4:
-          return kTsanRead4RuntimeEntry;
+          return kSanRead4RuntimeEntry;
         case 8:
-          return kTsanRead8RuntimeEntry;
+          return kSanRead8RuntimeEntry;
         case 16:
-          return kTsanRead16RuntimeEntry;
+          return kSanRead16RuntimeEntry;
         default:
           UNREACHABLE();
       }
@@ -2111,8 +2111,8 @@ void StoreIndexedInstr::EmitNativeCode(FlowGraphCompiler* compiler) {
       RepresentationUtils::RepresentationOfArrayElement(class_id());
   ASSERT(RequiredInputRepresentation(2) == Boxing::NativeRepresentation(rep));
 
-  if (!compiler->is_optimizing() && FLAG_target_thread_sanitizer) {
-    EmitTsanCallUnopt(compiler, this, [&]() -> const RuntimeEntry& {
+  if (!compiler->is_optimizing() && sanitize()) {
+    EmitSanCallUnopt(compiler, this, [&]() -> const RuntimeEntry& {
       if (index.IsRegister()) {
         __ ComputeElementAddressForRegIndex(R0, IsUntagged(), class_id(),
                                             index_scale(), index_unboxed_,
@@ -2124,15 +2124,15 @@ void StoreIndexedInstr::EmitNativeCode(FlowGraphCompiler* compiler) {
       }
       switch (RepresentationUtils::ValueSize(rep)) {
         case 1:
-          return kTsanWrite1RuntimeEntry;
+          return kSanWrite1RuntimeEntry;
         case 2:
-          return kTsanWrite2RuntimeEntry;
+          return kSanWrite2RuntimeEntry;
         case 4:
-          return kTsanWrite4RuntimeEntry;
+          return kSanWrite4RuntimeEntry;
         case 8:
-          return kTsanWrite8RuntimeEntry;
+          return kSanWrite8RuntimeEntry;
         case 16:
-          return kTsanWrite16RuntimeEntry;
+          return kSanWrite16RuntimeEntry;
         default:
           UNREACHABLE();
       }
@@ -2229,21 +2229,6 @@ void StoreIndexedInstr::EmitNativeCode(FlowGraphCompiler* compiler) {
     }
   } else {
     UNREACHABLE();
-  }
-
-  if (FLAG_target_memory_sanitizer) {
-    if (index.IsRegister()) {
-      __ ComputeElementAddressForRegIndex(TMP, IsUntagged(), class_id(),
-                                          index_scale(), index_unboxed_, array,
-                                          index.reg());
-    } else {
-      __ ComputeElementAddressForIntIndex(TMP, IsUntagged(), class_id(),
-                                          index_scale(), array,
-                                          Smi::Cast(index.constant()).Value());
-    }
-    const intptr_t length_in_bytes = RepresentationUtils::ValueSize(
-        RepresentationUtils::RepresentationOfArrayElement(class_id()));
-    __ MsanUnpoison(TMP, length_in_bytes);
   }
 }