+++ runtime/vm/compiler/backend/il_printer.cc
@@ -1253,7 +1253,7 @@ void TsanFuncEntryExitInstr::PrintOperandsTo(BaseTextBuffer* f) const {
   }
 }
 
-void TsanReadWriteInstr::PrintOperandsTo(BaseTextBuffer* f) const {
+void SanReadWriteInstr::PrintOperandsTo(BaseTextBuffer* f) const {
   instance()->PrintTo(f);
   f->Printf(" . %s", slot().Name());
   if (kind_ == kRead) {
@@ -1263,7 +1263,7 @@ void TsanReadWriteInstr::PrintOperandsTo(BaseTextBuffer* f) const {
   }
 }
 
-void TsanReadWriteIndexedInstr::PrintOperandsTo(BaseTextBuffer* f) const {
+void SanReadWriteIndexedInstr::PrintOperandsTo(BaseTextBuffer* f) const {
   array()->PrintTo(f);
   f->AddString(", ");
   index()->PrintTo(f);